Move all interactions with window.localStorage to separate module
